Finished, except for suggested products slider, not sure to implement yet or not :P
