+++ src/pages/ProductDetailsPage/ProductDetailsPage.scss
@@ -13,210 +13,6 @@
     }
 }
 
-.productCustomization {
-    width: 944px;
-    height: 464px;
-    margin-top: 40px;
-    display: flex;
-    gap: 16px;
-
-    @include onMedium {
-        width: 100%;
-    }
-
-    @include onSmall {
-        flex-direction: column;
-        height: 100%;
-        margin-top: 32px;
-    }
-
-    &-productImages {
-        @include flex-content(flex-start, column, center);
-        width: 80px;
-        min-width: 35px;
-        overflow: hidden;
-        gap: 16px;
-        order: 1;
-
-        @include onMedium {
-            gap: 8px;
-        }
-
-        @include onSmall {
-            order: 2;
-            flex-direction: row;
-            width: 100%;
-            height: 55px;
-            margin-bottom: 24px;
-        }
-
-        &-image {
-            max-height: 80px;
-            height: 100%;
-            width: 100%;
-            padding: 0.3em;
-            cursor: pointer;
-            border: 1px solid $elementsColor;
-
-            @include onMedium {
-                height: 12%;
-            }
-
-            @include onSmall {
-                max-height: 100%;
-                height: 100%;
-                min-width: 17.5%;
-                max-width: max-content;
-            }
-
-            img {
-                object-fit: contain;
-                height: 100%;
-                margin: auto;
-            }
-
-            &--active {
-                border: 1px solid $primaryColor;
-            }
-        }
-    }
-
-    &-imageSelectedBox {
-        order: 2;
-        max-height: 464px;
-        max-width: 464px;
-        min-width: 287px;
-        width: 50%;
-        padding: 0.8em;
-        margin-right: 35px;
-
-        @include onMedium {
-            margin-right: 0;
-            padding: 0 2em;
-        }
-
-        @include onSmall {
-            width: 100%;
-            height: 288px;
-            margin: auto;
-            padding: 0 2em;
-            order: 1;
-        }
-
-        &-image {
-            margin: auto;
-            object-fit: contain;
-            height: 80%;
-
-            @include onSmall {
-                height: 100%;
-                max-height: 360px;
-                // order: 1;
-            }
-        }
-    }
-
-    &-specs {
-        min-width: 237px;
-        order: 3;
-        flex-grow: 1;
-
-        &-availableColors {
-            margin-bottom: 24px;
-        }
-
-        &-colors {
-            margin-top: 8px;
-            display: flex;
-            gap: 8px;
-
-            &-colorBox {
-                width: 30px;
-                height: 30px;
-                border: 1px solid $elementsColor;
-                border-radius: 36px;
-                cursor: pointer;
-
-                :hover {
-                    border: 1px solid $primaryColor;
-                }
-
-                &--active {
-                    border: 1px solid $primaryColor;
-                }
-
-                &-color {
-                    border: 2px solid white;
-                    height: 100%;
-                    width: 100%;
-                    border-radius: inherit;
-                }
-            }
-        }
-
-        &-capacitySelection {
-            margin-bottom: 24px;
-
-            &-options {
-                display: flex;
-                gap: 8px;
-                margin-top: 8px;
-
-                &-option {
-                    font-weight: 500;
-                    border: 1px solid $iconsColor;
-
-                    &--active {
-                        background-color: $primaryColor;
-                        color: white;
-                    }
-                }
-            }
-        }
-
-        &-priceBox {
-            display: flex;
-            align-items: center;
-            gap: 8px;
-            margin-bottom: 16px;
-
-            &-priceDiscount {
-                font-size: 22px;
-                line-height: 28px;
-                text-decoration-line: line-through;
-                color: $secondaryColor;
-            }
-        }
-
-        &-actionButtons {
-            display: flex;
-            align-items: center;
-            gap: 8px;
-            height: 48px;
-            margin-bottom: 32px;
-
-            &-addToCart {
-                background-color: $primaryColor;
-                flex-grow: 1;
-                height: 100%;
-                color: white;
-            }
-
-            &-like {
-                width: 48px;
-                height: 100%;
-                border: 1px solid $iconsColor;
-            }
-        }
-
-        &-techs {
-            display: flex;
-            flex-direction: column;
-            gap: 8px;
-        }
-    }
-}
-
 .techSpecs {
     @include flex-content(space-between, row, center);
     gap: 50px;
@@ -226,72 +22,8 @@
     }
 }
 
-.description {
-    width: 1136px;
-    display: flex;
-    gap: 64px;
-    margin-top: 56px;
-    margin-bottom: 80px;
-
-    @include onLarge {
-        margin-top: 80px;
-    }
-
-    @include onMedium {
-        width: 100%;
-        margin-top: 64px;
-        margin-bottom: 64px;
-        flex-direction: column;
-    }
-
-    @include onSmall {
-        margin-top: 56px;
-        margin-bottom: 56px;
-    }
-
-    &-aboutBox {
-        flex-grow: 1;
-        flex-basis: 34.5%;
-
-        @include onMedium {
-            width: 100%;
-        }
-
-        &-section {
-            margin-top: 32px;
-
-            &-text {
-                margin-top: 16px;
-            }
-        }
-    }
-
-    &-techSpecsBox {
-        flex-grow: 1;
-
-        &-items {
-            display: flex;
-            flex-direction: column;
-            gap: 8px;
-            margin-top: 26px;
-        }
-    }
-}
-
 .divider {
     border: 1px solid $elementsColor;
-}
-
-.colors-divider {
-    margin-bottom: 24px;
-}
-
-.capacity-divider {
-    margin-bottom: 32px;
-}
-
-.descriptionDivider {
-    margin-top: 16px;
 }
 
 p {
