+++ src/pages/ProductDetailsPage/components/ProductDescription/ProductDescription.scss
@@ -1,0 +1,57 @@
+@import '../../../../styles/utils';
+
+.description {
+    width: 1136px;
+    display: flex;
+    gap: 64px;
+    margin-top: 56px;
+    margin-bottom: 80px;
+
+    @include onLarge {
+        margin-top: 80px;
+    }
+
+    @include onMedium {
+        width: 100%;
+        margin-top: 64px;
+        margin-bottom: 64px;
+        flex-direction: column;
+    }
+
+    @include onSmall {
+        margin-top: 56px;
+        margin-bottom: 56px;
+    }
+
+    &-aboutBox {
+        flex-grow: 1;
+        flex-basis: 34.5%;
+
+        @include onMedium {
+            width: 100%;
+        }
+
+        &-section {
+            margin-top: 32px;
+
+            &-text {
+                margin-top: 16px;
+            }
+        }
+    }
+
+    &-techSpecsBox {
+        flex-grow: 1;
+
+        &-items {
+            display: flex;
+            flex-direction: column;
+            gap: 8px;
+            margin-top: 26px;
+        }
+    }
+}
+
+.descriptionDivider {
+    margin-top: 16px;
+}
