+++ src/pages/ProductDetailsPage/components/Tech.tsx
