+++ src/pages/ProductDetailsPage/components/index.ts
@@ -1,0 +1,3 @@
+export {default as ProductCustomization } from './ProductCustomization/ProductCustomization';
+export {default as ProductDescription } from './ProductDescription/ProductDescription';
+export {default as Tech } from './Tech';
